Finishing up the C API.
There's still a bit more polishing work to do (high level docs and
examples), but I think the core bits are done.

This resulted in a new addition to the Rust API: a RegexBuilder. I've
wanted this for a while as a way to tweak limits and set default flag
values.